+++ src/commands/thermal_resistances/thermal_resistance_25.rs
@@ -1,8 +1,7 @@
 use crate::commands::command_map::Command;
 use crate::{impl_two_byte_read_command, impl_two_byte_write_command};
+use crate::units::{ElectricalResistance, ohm};
 use bitbybit::bitfield;
-use uom::si::electrical_resistance::ohm;
-use uom::si::f32::ElectricalResistance;
 
 /// This command is used to read and write the thermal resistance value at 25 degrees Celsius.
 /// It is represented as a 16-bit unsigned integer in ohms (Ω). The default value is set to 0x2710,
@@ -31,7 +30,7 @@ pub struct ThermalResistance25 {
 impl ThermalResistance25 {
     /// Returns the thermal resistance value in ohms at 25 degrees Celsius.
     pub fn thermal_resistance(&self) -> ElectricalResistance {
-        ElectricalResistance::new::<ohm>(self.raw_thermal_resistance() as f32)
+        ElectricalResistance::new::<ohm>(u32::from(self.raw_thermal_resistance()))
     }
 }
 impl_two_byte_read_command!(ThermalResistance25, Command::ThermalResistance25);
